Rom Op Overloads; Xml Class Introduction
Overloaded some equivalency operators for the Rom class. This was done
as the new Xml class contains a series of Roms and needs to be able to
sort them.

Xml class will extract features from a data file in XML format and turn
it into a Rom class instances kept in a vector. This is done to allow
for easy locating of a Rom's verification data.